pyproject.toml
[project]
name = "storepilotai"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "fastapi>=0.118.0",
    "openai-agents>=0.2.4",
    "pydantic>=2.11.7",
    "uvicorn>=0.35.0",
]
